final ai project/requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
werkzeug==2.0.1
openai==0.27.0
python-dotenv==0.19.0
requests==2.31.0 